requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-dotenv
pydantic
cryptography